refactor(agent): extract the queued-request codec from AgentRuntime
Turning an AgentRunRequest into the JSON stored on a queued run row, and
reading it back, moves into AgentRunRequestCodec.

The two directions are a single contract: whatever dehydrate() writes,
rehydrate() has to be able to read back with the same meaning — including
the planned cost and idempotency key that ToolOptions::toArray() drops,
which is why the codec carries them out of band. Stating that contract in
one class lets it be tested directly instead of inferred from a run, and
takes the first of the seams the roadmap lists for the runtime.

AgentRuntime loses its SkillRepository and PromptSnippetRepository
constructor arguments; only the serialisation used them, and the codec
takes them instead. Both existing construction sites pass fewer positional
arguments than that, so neither is affected, and consumers go through
AgentRuntimeInterface. The codec argument is optional for the same reason
the collaborators above it are: the bare test wiring constructs the
runtime positionally, and a null falls back to a codec over the
configuration repository the runtime already holds — the same absence of
forced skills and snippets that wiring had before.
